refactor :: [#41] 참여 팀 Presentation 계층 오타 수정
