+++ src/team/presentation/team.presentation.module.ts
@@ -1,13 +1,13 @@
 import { Module } from '@nestjs/common';
 
 import { TeamApplicationModule } from '../application/team.application.module';
-import { TeamWrtieAdapter } from './team.write.adapter';
 import { RequestTeamFormMapper } from './form/request/request.team.mapper';
 import { TeamReadAdapter } from './team.read.adapter';
+import { TeamWriteAdapter } from './team.write.adapter';
 
 @Module({
     imports: [TeamApplicationModule],
-    controllers: [TeamReadAdapter, TeamWrtieAdapter],
+    controllers: [TeamReadAdapter, TeamWriteAdapter],
     providers: [RequestTeamFormMapper]
 })
 export class TeamPresentationModule {}
