+++ src/team/presentation/team.write.adapter.ts
@@ -1,6 +1,6 @@
 import { Body, Controller, Get, Inject, Param, Post, Res, UseGuards } from '@nestjs/common';
 import { TeamWriteService } from '../application/team.write.service';
-import { RequestTeamForm } from './form/request/reqeust.team';
+import { RequestTeamForm } from './form/request/request.team';
 import { RequestTeamFormMapper } from './form/request/request.team.mapper';
 import { RequestTeamOrderForm } from './form/request/request.team.order';
 import { JwtAuthGuard } from 'src/common/guard/jwt.guard';
@@ -10,7 +10,7 @@ import { UserID } from 'src/common/decorator/user.decorator';
 import { ResponseTeamTokenForm } from './form/response/response.team';
 
 @Controller('/team')
-export class TeamWrtieAdapter {
+export class TeamWriteAdapter {
     constructor(
         @Inject('write_impl')
         private readonly writeService: TeamWriteService,
